constraint list for playground
