+++ frontend/src/component/playground/Playground/PlaygroundResultsTable/FeatureResultInfoPopoverCell/FeatureStrategyList/StrategyList/StrategyItem/StrategyExecution/ConstraintExecution/ConstraintError/LegacyConstraintError.tsx
@@ -8,6 +8,7 @@ import type {
 const StyledConstraintErrorDiv = styled('div')(({ theme }) => ({
     display: 'flex',
     alignItems: 'center',
+    marginTop: theme.spacing(1),
     color: theme.palette.error.main,
 }));
