+++ frontend/src/component/playground/Playground/PlaygroundResultsTable/FeatureResultInfoPopoverCell/FeatureStrategyList/StrategyList/StrategyItem/StrategyExecution/ConstraintExecution/ConstraintExecution.tsx
@@ -1,59 +1,31 @@
-import { Fragment, type VFC } from 'react';
+import type { FC } from 'react';
 import type {
     PlaygroundConstraintSchema,
     PlaygroundRequestSchema,
 } from 'openapi';
-import { objectId } from 'utils/objectId';
-import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
-import { StrategySeparator } from 'component/common/StrategySeparator/LegacyStrategySeparator';
-import { styled } from '@mui/material';
-import { ConstraintAccordionView } from 'component/common/ConstraintAccordion/ConstraintAccordionView/ConstraintAccordionView';
 import { ConstraintError } from './ConstraintError/ConstraintError';
-import { ConstraintOk } from './ConstraintOk/ConstraintOk';
+import { ConstraintOk } from './ConstraintOk/ConstraintOk'; // TODO: ask if this is needed
+import { ConstraintItem } from 'component/common/ConstraintsList/ConstraintItem/ConstraintItem';
 
 interface IConstraintExecutionProps {
-    constraints?: PlaygroundConstraintSchema[];
+    constraint?: PlaygroundConstraintSchema;
     input?: PlaygroundRequestSchema;
 }
 
-export const ConstraintExecutionWrapper = styled('div')(() => ({
-    width: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-}));
-
-export const ConstraintExecution: VFC<IConstraintExecutionProps> = ({
-    constraints,
+export const ConstraintExecution: FC<IConstraintExecutionProps> = ({
+    constraint,
     input,
 }) => {
-    if (!constraints) return null;
+    if (!constraint) return null;
 
     return (
-        <ConstraintExecutionWrapper>
-            {constraints?.map((constraint, index) => (
-                <Fragment key={objectId(constraint)}>
-                    <ConditionallyRender
-                        condition={index > 0}
-                        show={<StrategySeparator text='AND' />}
-                    />
-                    <ConstraintAccordionView
-                        constraint={constraint}
-                        compact
-                        renderAfter={
-                            <ConditionallyRender
-                                condition={constraint.result}
-                                show={<ConstraintOk />}
-                                elseShow={
-                                    <ConstraintError
-                                        input={input}
-                                        constraint={constraint}
-                                    />
-                                }
-                            />
-                        }
-                    />
-                </Fragment>
-            ))}
-        </ConstraintExecutionWrapper>
+        <>
+            <ConstraintItem {...constraint} />
+            {constraint.result ? (
+                <ConstraintOk />
+            ) : (
+                <ConstraintError input={input} constraint={constraint} />
+            )}
+        </>
     );
 };
